+++ src/main/java/com/leets/commitatobe/domain/auth/dto/GitHubDto.java
@@ -1,0 +1,6 @@
+package com.leets.commitatobe.domain.auth.dto;
+
+public record GitHubDto(
+	String userId
+) {
+}
