+++ src/main/java/com/leets/commitatobe/domain/auth/service/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package com.leets.commitatobe.domain.login.service;
+package com.leets.commitatobe.domain.auth.service;
 
 import java.util.Collections;
 
@@ -8,7 +8,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.leets.commitatobe.domain.login.domain.CustomUserDetails;
+import com.leets.commitatobe.domain.auth.domain.CustomUserDetails;
 import com.leets.commitatobe.domain.user.domain.User;
 import com.leets.commitatobe.domain.user.repository.UserRepository;
