+++ src/main/java/com/leets/commitatobe/global/jwt/provider/JwtProvider.java
@@ -13,7 +13,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Component;
 
-import com.leets.commitatobe.domain.login.domain.CustomUserDetails;
+import com.leets.commitatobe.domain.auth.domain.CustomUserDetails;
 import com.leets.commitatobe.global.jwt.dto.JwtResponse;
 
 import io.jsonwebtoken.Claims;
